# swagger.yaml
openapi: 3.0.0
info:
  title: Invoice backend
  description:  Invoice API docs
  version: 1.0.0
paths:
...
  /client:
    post:
      tags:
      - "client"
      description: "Create clients"
      requestBody:
       description: Schema of client
       required: true
       content:
        application/json:
         schema:
          $ref: '#/components/schemas/Client'
      responses:
        "200":
          description: Success Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: true
                 code:
                  type: "integer"
                  example: 201
                 message:
                   type: "string"
                   example: request successful client created succesfully
               data:
                type: "object"
                properties:
                    _id: 
                     type: "string"
                     example: "6200a70e9fec2c68f0cdc6de"
                    client_name:
                     type: "string"
                     example: "Harshith"
                    billing_address:
                     type: "string"
                     example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                     Nagin street, Sehari, Guntur Dt, AP, 524560
                     8130657523
                     example@gmail.com
                     PAN: ADIBC9018B
                     GSTIN: 37AAIBB8121B1Z1"
                    shipping_address:
                     type: "string"
                     example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                     Nagin street, Sehari, Guntur Dt, AP, 524560
                     8130657523
                     example@gmail.com
                     PAN: ADIBC9018B
                     GSTIN: 37AAIBB8121B1Z1"
                    date_of_contract:
                     type: "string"
                     example: "1/7/2021"
                    payment_terms:
                     type: "string"
                     example: "payment terms"
                    notes:
                     type: "string"
                     example: "Notes"
                    terms:
                     type: "string"
                     example: "Terms"
                    __v:
                     type: "number"
                     example: 0
                    
        "400":
          description: Bad request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 400
                 message:
                   type: "string"
                   example: Please enter _____       
        "500":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 500
                 message:
                   type: "string"
                   example: error message
                        
    get:
      tags:
      - "client"
      description: "Get All Clients"
      responses:
        "200":
          description: Success Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: true
                 code:
                  type: "integer"
                  example: 200
                 message:
                   type: "string"
                   example: request successful 
               data:
                type: "object"
                properties:
                 results :
                  type: "array"
                  items:
                   type: "object"
                   properties:
                    _id: 
                      type: "string"
                      example: "6200a70e9fec2c68f0cdc6de"
                    client_name:
                     type: "string"
                     example: "Harshith"
                    billing_address:
                     type: "string"
                     example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                     Nagin street, Sehari, Guntur Dt, AP, 524560
                     8130657523
                     example@gmail.com
                     PAN: ADIBC9018B
                     GSTIN: 37AAIBB8121B1Z1"
                    shipping_address:
                     type: "string"
                     example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                     Nagin street, Sehari, Guntur Dt, AP, 524560
                     8130657523
                     example@gmail.com
                     PAN: ADIBC9018B
                     GSTIN: 37AAIBB8121B1Z1"
                    date_of_contract:
                     type: "string"
                     example: "1/7/2021"
                    payment_terms:
                     type: "string"
                     example: "payment terms"
                    notes:
                     type: "string"
                     example: "Terms"
                    terms:
                     type: "string"
                     example: "Terms"
                    __v:
                     type: "number"
                     example: 0
        "404":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 404
                 message:
                   type: "string"
                   example: not found
        "500":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 500
                 message:
                   type: "string"
                   example: error message
  /client/{client_id}:
    patch:
      tags:
      - "client"
      description: "Updates a Client"
      parameters:
      - name: client_id
        required: true
        in: path
        schema:
          type: integer
      requestBody:
       description: Schema of PATCH Request
       required: true
       content:
        application/json:
         schema:
          $ref: '#/components/schemas/Client'
      responses:
        "200":
          description: Success Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: true
                 code:
                  type: "integer"
                  example: 204
                 message:
                   type: "string"
                   example: request successful client updated successfully
               data:
                type: "object"
                properties:
                 _id: 
                  type: "string"
                  example: "6200a70e9fec2c68f0cdc6de"
                 client_name:
                  type: "string"
                  example: "Harshith"
                 billing_address:
                  type: "string"
                  example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                  Nagin street, Sehari, Guntur Dt, AP, 524560
                  8130657523
                  example@gmail.com
                  PAN: ADIBC9018B
                  GSTIN: 37AAIBB8121B1Z1"
                 shipping_address:
                  type: "string"
                  example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                  Nagin street, Sehari, Guntur Dt, AP, 524560
                  8130657523
                  example@gmail.com
                  PAN: ADIBC9018B
                  GSTIN: 37AAIBB8121B1Z1"
                 date_of_contract:
                  type: "string"
                  example: "1/7/2021"
                 payment_terms:
                  type: "string"
                  example: "payment terms"
                 notes:
                  type: "string"
                  example: "Notes"
                 terms:
                  type: "string"
                  example: "Terms"
                 __v:
                  type: "number"
                  example: 0
        "400":
          description: Bad request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 400
                 message:
                   type: "string"
                   example: Please enter ____
        "404":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 404
                 message:
                   type: "string"
                   example: client not found
        "500":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 500
                 message:
                   type: "string"
                   example: error message
    get:
      tags:
      - "client"
      description: "Get Client by id"
      parameters:
      - name: client_id
        required: true
        in: path
        schema:
          type: integer
      responses:
        "200":
          description: Success Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: true
                 code:
                  type: "integer"
                  example: 200
                 message:
                   type: "string"
                   example: request successful
               data:
                type: "object"
                properties:
                    _id: 
                     type: "string"
                     example: "6200a70e9fec2c68f0cdc6de"
                    client_name:
                     type: "string"
                     example: "Harshith"
                    billing_address:
                     type: "string"
                     example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                     Nagin street, Sehari, Guntur Dt, AP, 524560
                     8130657523
                     example@gmail.com
                     PAN: ADIBC9018B
                     GSTIN: 37AAIBB8121B1Z1"
                    shipping_address:
                     type: "string"
                     example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                     Nagin street, Sehari, Guntur Dt, AP, 524560
                     8130657523
                     example@gmail.com
                     PAN: ADIBC9018B
                     GSTIN: 37AAIBB8121B1Z1"
                    date_of_contract:
                     type: "string"
                     example: "1/7/2021"
                    payment_terms:
                     type: "string"
                     example: "payment terms"
                    notes:
                     type: "string"
                     example: "Notes"
                    terms:
                     type: "string"
                     example: "Terms"
                    __v:
                     type: "number"
                     example: 0
        "404":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 404
                 message:
                   type: "string"
                   example: client not found       
        "500":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 500
                 message:
                   type: "string"
                   example: error message  
          
    delete:
      tags:
      - "client"
      description: "Delete Client"
      parameters:
      - name: client_id
        required: true
        in: path
        schema:
          type: integer
      responses:
        "200":
          description: Success Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: true
                 code:
                  type: "integer"
                  example: 204
                 message:
                   type: "string"
                   example: request successful client deleted successfully
        "404":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 404
                 message:
                   type: "string"
                   example: client not found 
        "500":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 500
                 message:
                   type: "string"
                   example: error message       
components:
  schemas:
...
    Client:
     type: "object"
     required:
     - "client_name"
     - "billing_address"
     - "date_of_contract"
     - "payment_terms"
     properties:
      client_name:
        type: "string"
        example: "Harshith"
      billing_address:
        type: "string"
        example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
        Nagin street, Sehari, Guntur Dt, AP, 524560
        8130657523
        example@gmail.com
        PAN: ADIBC9018B
        GSTIN: 37AAIBB8121B1Z1"
      shipping_address:
        type: "string"
        example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
        Nagin street, Sehari, Guntur Dt, AP, 524560
        8130657523
        example@gmail.com
        PAN: ADIBC9018B
        GSTIN: 37AAIBB8121B1Z1"
      date_of_contract:
        type: "string"
        example: "1/7/2021"
      payment_terms:
        type: "string"
        example: "payment terms"
      notes:
        type: "string"
        example: "Terms"
      terms:
        type: "string"
        example: "Terms"
...